// openapi/vitrivr-engine/model/queryResult.ts
/**
 * vitrivr engine API
 * Rest API for the vitrivr engine project. Provides query (runtime) and extraction (ingestion) endpoints
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { QueryResultRetrievable } from './queryResultRetrievable';


export interface QueryResult { 
    retrievables: Array<QueryResultRetrievable>;
}